+++ Enamel/Systems/PlayerButtonsSystem.cs
@@ -51,7 +51,7 @@ public class PlayerButtonsSystem : MoonTools.ECS.System
             var spellCard = _world.CreateEntity();
             Set(spellCard, new PositionComponent(screenX, 320));
             Set(spellCard, new DimensionsComponent(30, 30));
-            Set(spellCard, new TextureIndexComponent((int) Sprite.YellowSquare));
+            Set(spellCard, new TextureIndexComponent(Sprite.YellowSquare));
             Set(spellCard, new TextComponent(TextStorage.GetId(spellIdComponent.SpellId.ToName()), Constants.SPELL_CARD_TEXT_SIZE, Constants.SpellCardTextColour));
             Set(spellCard, new OnClickComponent(ClickEvent.PrepSpell));
             Set(spellCard, new SpellToPrepOnClickComponent(spellIdComponent.SpellId));
